Move middleware out of runserver
